+++ CHANGELOG.md
@@ -7,6 +7,21 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [7.1.0] - 2024-05-24
+
+- Adds implementation for a new method `getConfigFieldsInfo` to fetch the plugin config fields.
+- Adds `null` state for `firstFactors` and `providers` by adding `is_first_factors_null` and `is_third_party_providers_null` fields in `tenant_configs` table
+
+### Migration
+
+```sql
+ALTER TABLE tenant_configs ADD COLUMN is_first_factors_null BOOLEAN DEFAULT TRUE;
+ALTER TABLE tenant_configs ADD COLUMN is_third_party_providers_null BOOLEAN DEFAULT TRUE;
+
+ALTER TABLE tenant_configs ALTER COLUMN is_first_factors_null DROP DEFAULT;
+ALTER TABLE tenant_configs ALTER COLUMN is_third_party_providers_null DROP DEFAULT;
+```
+
 ## [7.0.1] - 2024-04-17
 
 - Fixes issues with partial failures during tenant creation
@@ -208,6 +223,7 @@ CREATE INDEX app_id_to_user_id_user_id_index ON app_id_to_user_id (user_id);
 
     ```sql
     -- helper stored procedures
+    DELIMITER //
 
     CREATE PROCEDURE st_drop_all_fkeys()
     BEGIN
@@ -264,7 +280,7 @@ CREATE INDEX app_id_to_user_id_user_id_index ON app_id_to_user_id (user_id);
         END LOOP;
 
         CLOSE dropCur;
-    END
+    END //
 
     --
 
@@ -323,7 +339,7 @@ CREATE INDEX app_id_to_user_id_user_id_index ON app_id_to_user_id (user_id);
         END LOOP;
 
         CLOSE dropCur;
-    END
+    END //
 
     --
 
@@ -382,7 +398,7 @@ CREATE INDEX app_id_to_user_id_user_id_index ON app_id_to_user_id (user_id);
         END LOOP;
 
         CLOSE dropCur;
-    END
+    END //
 
     --
 
@@ -441,7 +457,7 @@ CREATE INDEX app_id_to_user_id_user_id_index ON app_id_to_user_id (user_id);
         END LOOP;
 
         CLOSE dropCur;
-    END
+    END //
 
     --
 
@@ -473,8 +489,9 @@ CREATE INDEX app_id_to_user_id_user_id_index ON app_id_to_user_id (user_id);
         execute add_column_sql;
           SELECT 'Column Successfully  Created!' INTO p_status_message;
         END IF;
-    END
+    END //
 
+    DELIMITER ;
     -- Drop constraints and indexes
 
     CALL st_drop_all_fkeys();